LPS-60609 Move exceptions
